configure initial info on the character cell
